Archivos para servidor

+++ database/migrations/2020_04_13_063929_create_productsorders_table.php
@@ -21,7 +21,7 @@ class CreateProductsordersTable extends Migration
      
             $table->integer('unidades');
             $table->float('precio_unidad', 6, 2);
-            $table->integer('total');
+            $table->float('total', 6, 2);
             
 
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
